requirements.txt
pydantic~=2.0
openai~=1.29
numpy~=1.26
anthropic>=0.29.0
google-generativeai~=0.7
